#216961 Add custom default-sorting to specific CLP (#515)
